refactor(pki): extract appchain communication functions to separate file
